[Partial] Socket to check all service connections
